ファイルの存在確認

+++ app/Http/Controllers/HelloController.php
@@ -79,12 +79,19 @@ class HelloController extends Controller
         // シンボリックリンク /storage/app/public/から、先頭に追加
         // Storage::disk('public')->prepend($this->fname, $msg);
 
+
         // /storage/app/public/hello.txtがあれば先に削除
-        Storage::disk('public')->delete('bk_' . $this->fname);
+        if(Storage::disk('public')->exists('bk_'.$this->fname)){
+            Storage::disk('public')->delete('bk_' . $this->fname);
+        }
+
         // /storage/app/public/hello.txtをコピーして、bk_hello.txtを作る
         Storage::disk('public')->copy($this->fname, 'bk_' . $this->fname);
+
         // /storage/app/内に、bk_hello.txtがあれば削除
-        Storage::disk('local')->delete('bk_' . $this->fname);
+        if (Storage::disk('local')->exists('bk_'.$this->fname)){
+            Storage::disk('local')->delete('bk_' . $this->fname);
+        }
         // /storage/app/から/storage/app/public/に移動
         Storage::disk('local')->move('public/bk_' . $this->fname, 'bk_' . $this->fname);
